apply dtos and optimizations for frontend
