+++ frontend/src/api/gradesProcessing.ts
@@ -21,7 +21,7 @@ export function calcProjectGrade(sheet: GradeSheetData) {
 const MAX_ROLE_POINTS = 10
 export function calcUserProjectGrade(sheet: GradeSheetData, userId: string) {
   const [projectPoints, maxPoints] = calcProjectGrade(sheet)
-  const user = sheet.participants.find((user) => user.participantID === userId)
+  const user = sheet.participants.find((user) => user.id === userId)
   if (!(user && user.engagement && user.rolePoints)) return null
   const points = (projectPoints * user.engagement) / 100 + user.rolePoints
   const max = maxPoints + MAX_ROLE_POINTS
