+++ frontend/src/components/GradeSheetInfo/index.ts
@@ -1,0 +1,1 @@
+export * from './GradeSheetInfo'
