+++ frontend/src/components/GradeSheetMentorGrades/index.ts
@@ -1,0 +1,1 @@
+export * from './GradeSheetMentorGrades'
