+++ frontend/src/components/GradeSheetParticipants/index.ts
@@ -1,0 +1,1 @@
+export * from './GradeSheetParticipants'
