+++ frontend/src/hooks/useQuery/useSheets.ts
@@ -1,12 +1,15 @@
 import { genericSearch, genericSort } from './useGenericQuery'
 import * as api from '../../api/Sheet.api'
-import { GradeSheet } from '../../models'
+import { GradeSheetDto } from '../../models'
 import { useQuery } from 'react-query'
 
 const queryKey = 'sheets'
 
-const useSheets = () => useQuery(queryKey, () => api.getSheets())
+const useSheets = (userMeData: any) =>
+  useQuery(queryKey, () => api.getSheets(userMeData.type), {
+    enabled: !!userMeData,
+  })
 export default useSheets
 
-export const searchSheet = genericSearch<GradeSheet>(queryKey)
-export const sortSheets = genericSort<GradeSheet>(queryKey)
+export const searchSheet = genericSearch<GradeSheetDto>(queryKey)
+export const sortSheets = genericSort<GradeSheetDto>(queryKey)
